rename: 쿠폰 관련 컴포넌트 Coupon 폴더로 이동
