+++ src/pages/CouponPage.tsx
@@ -1,5 +1,5 @@
-import CouponHeader from "@/components/common/Header/CouponHeader";
-import CouponInfoTable from "@/components/common/Table/CouponInfoTable";
+import CouponHeader from "@/components/Coupon/CouponHeader";
+import CouponInfoTable from "@/components/Coupon/CouponInfoTable";
 import Title from "@/components/common/Title";
 import CouponContextProvider from "@/contexts/CouponContext";
